pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "peer-learning-factory"
version = "0.1.0"
description = "Agent orchestration system for generating peer learning content from a codebase"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.2",
    "langchain-core>=0.3",
    "pydantic>=2.0",
    "rich>=13.0",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.0",
    "anthropic>=0.40",
    "openai>=1.50",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
peer-factory = "src.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
